Update the service account relation with respect to Kubeflow integration

+++ tests/unit/test_spark_service_account.py
@@ -2,6 +2,7 @@ import json
 from unittest import mock
 
 import pytest
+import yaml
 from lib.charms.data_platform_libs.v0.spark_service_account import (
     ServiceAccountGoneEvent,
     ServiceAccountGrantedEvent,
@@ -20,6 +21,7 @@ SERVICE_ACCOUNT = "default:user"
 RELATION_INTERFACE = "spark_service_account"
 RELATION_NAME = "spark-service-account"
 SPARK_PROPS = {"foo": "bar"}
+RESOURCE_MANIFEST = {"foo": "bar"}
 
 
 class SparkServiceAccountProviderCharm(CharmBase):
@@ -39,6 +41,10 @@ class SparkServiceAccountProviderCharm(CharmBase):
         self.framework.observe(self.provider.on.account_released, self._on_account_released)
         self.framework.observe(self.on.add_config_action, self._on_add_config_action)
 
+    def _generate_service_account_manifest(self, service_account: str) -> dict:
+        print(f"Generated service account resource manifest: {service_account}")
+        return RESOURCE_MANIFEST
+
     def _create_service_account(self, service_account: str) -> None:
         print(f"Created service account: {service_account}")
 
@@ -49,15 +55,22 @@ class SparkServiceAccountProviderCharm(CharmBase):
         if not event.service_account:
             return
         service_account = event.service_account
-        self._create_service_account(service_account)
+        if not event.skip_creation:
+            self._create_service_account(service_account)
+        manifest = self._generate_service_account_manifest(service_account=service_account)
         self.provider.set_service_account(event.relation.id, service_account)
         self.provider.set_spark_properties(event.relation.id, json.dumps(SPARK_PROPS))
+        self.provider.set_resource_manifest(
+            event.relation.id, resource_manifest=yaml.dump(manifest)
+        )
 
     def _on_account_released(self, event: ServiceAccountReleasedEvent) -> None:
         if not event.service_account:
             return
         service_account = event.service_account
-        self._delete_service_account(service_account)
+        skip_creation = event.skip_creation
+        if not skip_creation:
+            self._delete_service_account(service_account)
 
     def _on_add_config_action(self, event: ActionEvent) -> None:
         conf = event.params["conf"]
@@ -72,10 +85,16 @@ class SparkServiceAccountRequirerCharm(CharmBase):
 
     META = {"name": REQUIRER_APP, "requires": {RELATION_NAME: {"interface": RELATION_INTERFACE}}}
 
+    CONFIG = {"options": {"skip-creation": {"type": "string", "default": "false"}}}
+
     def __init__(self, *args):
         super().__init__(*args)
+        skip_creation = self.config.get("skip-creation", "false") == "true"
         self.requirer = SparkServiceAccountRequirer(
-            self, relation_name=RELATION_NAME, service_account=SERVICE_ACCOUNT
+            self,
+            relation_name=RELATION_NAME,
+            service_account=SERVICE_ACCOUNT,
+            skip_creation=skip_creation,
         )
         self.framework.observe(self.requirer.on.account_granted, self._on_account_granted)
         self.framework.observe(self.requirer.on.account_gone, self._on_account_gone)
@@ -125,7 +144,7 @@ class TestSparkServiceAccountProvider:
         )
 
     @mock.patch.object(SparkServiceAccountProviderCharm, "_create_service_account")
-    def test_service_account_created(self, mock_create_sa):
+    def test_service_account_created_by_provider(self, mock_create_sa):
         relation = self.get_relation()
         state1 = State(relations=[relation], leader=True)
         relation.remote_app_data.update(
@@ -140,6 +159,34 @@ class TestSparkServiceAccountProvider:
 
         local_app_data = state2.get_relation(relation.id).local_app_data
         assert local_app_data["service-account"] == SERVICE_ACCOUNT
+        assert local_app_data["resource-manifest"] == yaml.dump(RESOURCE_MANIFEST)
+
+        assert "secret-extra" in local_app_data
+
+        secret_id = local_app_data["secret-extra"]
+        secret_content = state2.get_secret(id=secret_id).latest_content
+        assert secret_content is not None
+        spark_properties = json.loads(secret_content["spark-properties"])
+        assert spark_properties == SPARK_PROPS
+
+    @mock.patch.object(SparkServiceAccountProviderCharm, "_create_service_account")
+    def test_service_account_creation_skipped(self, mock_create_sa):
+        relation = self.get_relation()
+        state1 = State(relations=[relation], leader=True)
+        relation.remote_app_data.update(
+            {
+                "service-account": SERVICE_ACCOUNT,
+                "requested-secrets": json.dumps(["spark-properties"]),
+                "skip-creation": "true",
+            }
+        )
+
+        state2 = self.context.run(self.context.on.relation_changed(relation), state1)
+        assert not mock_create_sa.called
+
+        local_app_data = state2.get_relation(relation.id).local_app_data
+        assert local_app_data["service-account"] == SERVICE_ACCOUNT
+        assert local_app_data["resource-manifest"] == yaml.dump(RESOURCE_MANIFEST)
         assert "secret-extra" in local_app_data
 
         secret_id = local_app_data["secret-extra"]
@@ -187,6 +234,20 @@ class TestSparkServiceAccountProvider:
         self.context.run(self.context.on.relation_broken(relation), state1)
         mock_delete_sa.assert_called_with(SERVICE_ACCOUNT)
 
+    @mock.patch.object(SparkServiceAccountProviderCharm, "_delete_service_account")
+    def test_service_account_released_skip_deletion(self, mock_delete_sa):
+        relation = self.get_relation()
+        relation.remote_app_data.update(
+            {
+                "service-account": SERVICE_ACCOUNT,
+                "requested-secrets": json.dumps(["spark-properties"]),
+                "skip-creation": "true",
+            }
+        )
+        state1 = State(relations=[relation], leader=True)
+        self.context.run(self.context.on.relation_broken(relation), state1)
+        assert not mock_delete_sa.called
+
 
 @pytest.mark.usefixtures("only_with_juju_secrets")
 class TestSparkServiceAccountRequirer:
@@ -206,7 +267,9 @@ class TestSparkServiceAccountRequirer:
     @property
     def context(self):
         return Context(
-            charm_type=SparkServiceAccountRequirerCharm, meta=SparkServiceAccountRequirerCharm.META
+            charm_type=SparkServiceAccountRequirerCharm,
+            meta=SparkServiceAccountRequirerCharm.META,
+            config=SparkServiceAccountRequirerCharm.CONFIG,
         )
 
     @mock.patch.object(SparkServiceAccountRequirerCharm, "_consume_service_account")
@@ -240,7 +303,12 @@ class TestSparkServiceAccountRequirer:
         relation.remote_app_data.update(
             {"service-account": SERVICE_ACCOUNT, "secret-extra": props_secret.id}
         )
-        state1 = State(relations=[relation], secrets=[props_secret], leader=True)
+        state1 = State(
+            relations=[relation],
+            secrets=[props_secret],
+            leader=True,
+            config={"skip-creation": "true"},
+        )
 
         self.context.run(self.context.on.secret_changed(props_secret), state1)
 
@@ -248,6 +316,8 @@ class TestSparkServiceAccountRequirer:
         service_account, spark_properties = args
         assert service_account == SERVICE_ACCOUNT
         assert "foo" in json.loads(spark_properties)
+
+        print(state1.config)
 
     @mock.patch.object(SparkServiceAccountRequirerCharm, "_consume_service_account")
     def test_service_account_gone(self, mock_consume_sa):
